Moved some more functions into a helper file.
